+++ semcore/date-picker/src/DateRangePicker.jsx
@@ -88,8 +88,8 @@ class DateRangePickerRoot extends RangePickerAbstract {
       locale,
       disabled,
       disabledErrorText,
-      size,
       getI18nText,
+      animationsDisabled,
     } = this.asProps;
     const value = visible
       ? [preselectedValue[0] || this.asProps.value[0], preselectedValue[1] || this.asProps.value[1]]
@@ -101,11 +101,11 @@ class DateRangePickerRoot extends RangePickerAbstract {
       onChange,
       onDisplayedPeriodChange,
       locale,
-      w: size === 'm' ? 270 : 300,
       disabledDates: disabled,
       disabledErrorText,
       children: () => <InputTrigger.DateRange />,
       getI18nText,
+      animationsDisabled,
     };
   }
 }