+++ semcore/date-picker/src/MonthRangePicker.jsx
@@ -97,8 +97,8 @@ class MonthRangePickerRoot extends RangePickerAbstract {
       locale,
       disabled,
       disabledErrorText,
-      size,
       getI18nText,
+      animationsDisabled,
     } = this.asProps;
     const value = visible
       ? [preselectedValue[0] || this.asProps.value[0], preselectedValue[1] || this.asProps.value[1]]
@@ -110,12 +110,12 @@ class MonthRangePickerRoot extends RangePickerAbstract {
       onChange,
       onDisplayedPeriodChange,
       locale,
-      w: size === 'm' ? 215 : 235,
       parts: dateParts,
       disabledDates: disabled,
       disabledErrorText,
       children: () => <InputTrigger.DateRange />,
       getI18nText,
+      animationsDisabled,
     };
   }