+++ website/docs/table-group/data-table/examples/download-status.tsx
@@ -22,6 +22,7 @@ const Demo = () => {
       </DataTable.Head>
       <SpinContainer loading={loading} style={{ overflow: 'initial' }}>
         <DataTable.Body />
+        <SpinContainer.Overlay />
       </SpinContainer>
     </DataTable>
   );