+++ semcore/input-number/CHANGELOG.md
@@ -8,6 +8,18 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - `Intl.NumberFormat` formatter for display value.
 
+## [5.22.2] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/input` [4.22.1 ~> 4.22.2], `@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [5.22.1] - 2024-04-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/input` [4.22.0 ~> 4.22.1], `@semcore/utils` [4.22.0 ~> 4.22.1], `@semcore/core` [2.19.0 ~> 2.19.1]).
+
 ## [5.22.0] - 2024-03-27
 
 ### Changed