+++ semcore/side-panel/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.29.2] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/animation` [2.21.1 ~> 2.21.2], `@semcore/flex-box` [5.21.1 ~> 5.21.2], `@semcore/portal` [3.21.1 ~> 3.21.2], `@semcore/outside-click` [3.19.1 ~> 3.19.2], `@semcore/icon` [4.30.1 ~> 4.30.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [3.29.1] - 2024-04-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.0 ~> 4.22.1], `@semcore/animation` [2.21.0 ~> 2.21.1], `@semcore/flex-box` [5.21.0 ~> 5.21.1], `@semcore/portal` [3.21.0 ~> 3.21.1], `@semcore/outside-click` [3.19.0 ~> 3.19.1], `@semcore/icon` [4.30.0 ~> 4.30.1], `@semcore/core` [2.19.0 ~> 2.19.1]).
+
 ## [3.29.0] - 2024-03-27
 
 ### Changed