+++ semcore/spin-container/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [7.22.0] - 2024-03-27
+
+### Changed
+
+- Allowed to use `SpinContainer.Overlay` without `SpinContainer.Content`.
+
+## [7.21.2] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/spin` [5.20.1 ~> 5.20.2], `@semcore/flex-box` [5.21.1 ~> 5.21.2], `@semcore/animation` [2.21.1 ~> 2.21.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [7.21.1] - 2024-04-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.0 ~> 4.22.1], `@semcore/spin` [5.20.0 ~> 5.20.1], `@semcore/flex-box` [5.21.0 ~> 5.21.1], `@semcore/animation` [2.21.0 ~> 2.21.1], `@semcore/core` [2.19.0 ~> 2.19.1]).
+
 ## [7.21.0] - 2024-03-27
 
 ### Changed