+++ semcore/tag/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.30.3] - 2024-04-11
+
+### Fixed
+
+- Focused tag was rendering focus ring even if it wasn't marked as interactive.
+
+## [5.30.2] - 2024-04-10
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.1 ~> 4.22.2], `@semcore/icon` [4.30.1 ~> 4.30.2], `@semcore/flex-box` [5.21.1 ~> 5.21.2], `@semcore/core` [2.19.1 ~> 2.19.2]).
+
+## [5.30.1] - 2024-04-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.22.0 ~> 4.22.1], `@semcore/icon` [4.30.0 ~> 4.30.1], `@semcore/flex-box` [5.21.0 ~> 5.21.1], `@semcore/core` [2.19.0 ~> 2.19.1]).
+
 ## [5.30.0] - 2024-03-27
 
 ### Changed